+++ common/src/main/java/dev/wuffs/squatgrow/config/SquatGrowConfig.java
@@ -1,100 +1,103 @@
 package dev.wuffs.squatgrow.config;
 
+import dev.nanite.library.core.config.Config;
+import dev.nanite.library.core.config.ConfigValueGroup;
+import dev.nanite.library.core.config.values.BooleanConfigValue;
+import dev.nanite.library.core.config.values.FloatConfigValue;
+import dev.nanite.library.core.config.values.IntConfigValue;
+import dev.nanite.library.core.config.values.ListConfigValue;
 import dev.wuffs.squatgrow.SquatGrow;
-import me.shedaniel.autoconfig.ConfigData;
-import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry;
-import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.Comment;
-import net.minecraft.tags.TagKey;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.block.Blocks;
 
-import java.beans.Transient;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
+import java.util.stream.Stream;
 
-@Config(name = SquatGrow.MOD_ID + "-common")
-public class SquatGrowConfig implements ConfigData {
-    @Comment("Enable debug logging")
-    public boolean debug = false;
+public interface SquatGrowConfig {
+    Config config = Config.commonConfig(SquatGrow.MOD_ID);
 
-    @Comment("Require hoe to allow growth, LEGACY, PLEASE SWITCH TO THE NEW SYSTEM, SEE REQUIREMENTS")
-    public boolean requireHoe = false;
+    BooleanConfigValue debug = config
+            .booleanValue("debug", false)
+            .comments("Enable debug logging");
 
-    @Comment("Hoe takes damage on growth, LEGACY, PLEASE SWITCH TO THE NEW SYSTEM, SEE REQUIREMENTS")
-    public boolean hoeTakesDamage = false;
+    BooleanConfigValue useWhitelist = config
+            .booleanValue("useWhitelist", false)
+            .comments("Use whitelist instead of blacklist, default false");
 
-    @Comment("Use whitelist instead of blacklist, default false")
-    public boolean useWhitelist = false;
+    IntConfigValue range = config
+            .intValue("range", 3)
+            .comments("Range of effect, warning: this can cause lag if set too high")
+            .min(0)
+            .max(16);
 
-    @Comment("Range of effect, warning: this can cause lag if set too high")
-    @ConfigEntry.BoundedDiscrete(min = 0, max = 16)
-    public int range = 3;
+    FloatConfigValue chance = config
+            .floatValue("chance", 0.5f)
+            .comments("Chance for a block to grow, between 0 and 1")
+            .min(0f)
+            .max(1f);
 
-    @Comment("Growth chance")
-    public float chance = 0.5f;
+    IntConfigValue randomTickMultiplier = config
+            .intValue("randomTickMultiplier", 4)
+            .comments("Random tick multiplier, this is the amount of times the mod will call the randomTick method on the block for each block in the range")
+            .min(1)
+            .max(16);
 
-    @Comment("Don't use! Use randomTickMultiplier instead, this is here for backwards compatibility")
-    @ConfigEntry.BoundedDiscrete(min = 1, max = 16)
-    public int sugarcaneMultiplier = 4;
+    // TODO: Fix casting issue in Nanite Library
+    ListConfigValue<String> ignoreList = (ListConfigValue<String>) config
+            .stringListValue("ignoreList", Stream.of(
+                    Blocks.GRASS_BLOCK,
+                    Blocks.SHORT_GRASS,
+                    Blocks.TALL_GRASS,
+                    Blocks.NETHERRACK,
+                    Blocks.WARPED_NYLIUM,
+                    Blocks.CRIMSON_NYLIUM
+            ).map(e -> e.builtInRegistryHolder().key().identifier().toString()).toList())
+            .comments("List of blocks to blacklist/whitelist from twerking, Tags can be used by using #minecraft:<tag_name> or #modid:<tag_name>");
 
-    @Comment("Random tick multiplier, this is the amount of times the mod will call the randomTick method on the block for each block in the range")
-    @ConfigEntry.BoundedDiscrete(min = 1, max = 16)
-    public int randomTickMultiplier = 4;
+    BooleanConfigValue allowAdventureTwerking = config
+            .booleanValue("allowAdventureTwerking", true)
+            .comments("Allow twerking in adventure mode, default true");
 
-    @Comment("List of blocks to blacklist/whitelist from twerking, Tags can be used by using #minecraft:<tag_name> or #modid:<tag_name>")
-    public List<String> ignoreList = new ArrayList<>(Arrays.asList(
-            "minecraft:grass_block",
-            "minecraft:grass",
-            "minecraft:short_grass",
-            "minecraft:tall_grass",
-            "minecraft:netherrack",
-            "minecraft:warped_nylium",
-            "minecraft:crimson_nylium"
-    ));
+    BooleanConfigValue enableMysticalCrops = config
+            .booleanValue("enableMysticalCrops", true)
+            .comments("Enable Mystical Crops growth support");
 
-    @Comment("Allow twerking in adventure mode, default true")
-    public boolean allowAdventureTwerking = true;
+    BooleanConfigValue enableAE2Accelerator = config
+            .booleanValue("enableAE2Accelerator", true)
+            .comments("Enable AE2 crystal growth support, only available if AE2 is present");
 
-    @Comment("Enable Mystical Crops")
-    public boolean enableMysticalCrops = true;
+    IntConfigValue ae2Multiplier = config
+            .intValue("ae2Multiplier", 4)
+            .comments("AE2 growth multiplier, only available if AE2 is present")
+            .min(1)
+            .max(16);
 
-    @Comment("Enable AE2 accelerator")
-    public boolean enableAE2Accelerator = true;
+    BooleanConfigValue enableDirtToGrass = config
+            .booleanValue("enableDirtToGrass", true)
+            .comments("When the player is holding a grass block in their offhand, they will be able to randomly convert dirt into grass");
 
-    @Comment("AE2 growth multiplier, only available if AE2 is present")
-    @ConfigEntry.BoundedDiscrete(min = 1, max = 16)
-    public int ae2Multiplier = 4;
+    ConfigValueGroup requirementsGroup = config
+            .group("requirements")
+            .comments("Requirements for growing");
 
-    @Comment("When the player is holding a grass block in their offhand, they will be able to randomly convert dirt into grass")
-    public boolean enableDirtToGrass = true;
+    // TODO: Fix casting issue in Nanite Library
+    ListConfigValue<String> requiredItem = (ListConfigValue<String>) requirementsGroup
+            .stringListValue("items", List.of())
+            .comments(
+                "List of items required to grow, leave empty to disable, can be either item ids or tags (use #minecraft:<tag_name> or #modid:<tag_name>)"
+            );
 
-    @ConfigEntry.Category("requirements")
-    @Comment("Requirements for growing")
-    public Requirements requirements = new Requirements();
+    // TODO: Fix casting issue in Nanite Library
+    ListConfigValue<String> requiredItemEnchants = (ListConfigValue<String>) requirementsGroup
+            .stringListValue("enchants", List.of())
+            .comments("List of enchantments required on the item to grow, leave empty to disable, format is <enchantment_id>@<level>, for example minecraft:fortune@3 or, optionally, omit the level to just check for the presence of the enchantment minecraft:fortune");
 
-    public static class Requirements {
-        @Comment("Enabled the new requirements system")
-        public boolean enabled = true;
+    IntConfigValue durabilityCost = requirementsGroup
+            .intValue("durabilityCost", 0)
+            .comments("Whether the required item should take durability damage when used to grow a block. If the item is not durable, this will have no effect. Leave as 0 to disable durability damage.");
 
-        @Comment("List of blocks that require a hoe to grow, leave empty to disable")
-        public List<String> heldItemRequirement = new ArrayList<>();
-
-        @Comment("Map of equipment slots to items required to grow, leave empty to disable")
-        public Map<EquipmentSlot, String> equipmentRequirement = Map.of();
-
-        @Comment("Durability based items take damage when used to grow")
-        public boolean requiredItemTakesDamage = false;
-
-        @Comment("Amount of damage to take when used to grow")
-        public int durabilityDamage = 1;
-
-        @Comment(
-                "Enchantment required to grow, leave empty to disable\n"
-        )
-        public String requiredEnchantment = "";
-    }
+    FloatConfigValue durabilityChance = requirementsGroup
+            .floatValue("durabilityChance", 1f)
+            .min(0f)
+            .max(1f)
+            .comments("The chance for the required item to take durability damage when used to grow a block, between 0 and 1. This is only used if durabilityCost is greater than 0.");
 }